6th commit

+++ src/main/java/com/aoao/service/impl/PatientServiceImpl.java
@@ -20,81 +20,99 @@ public class PatientServiceImpl implements IPatientService {
     @Autowired
     private IPatientDao patientDao;
 
+
     @Override
-    public List<String> findAllPatientNumber() {
+    public List<Integer> findAllPatientNumber() {
         return patientDao.findAllPatientNumber();
     }
 
     @Override
-    public PatientLv1 findPatientByPatientNumber(String patient_number_lv1) {
+    public PatientLv1 findPatientByPatientNumber(Integer patient_number_lv1) {
         PatientLv1 patient = patientDao.findPatientByPatientNumber(patient_number_lv1);
         return patient;
     }
 
     @Override
-    public void updatePatientLv1Info(PatientLv1 patientLv1) {
+    public String updatePatientLv1Info(PatientLv1 patientLv1) {
         if (patientLv1.getPatient_number_lv1() == 0){
-            System.out.println("病人一級表更新時未傳入patient_number_lv1");
-            return;
+            return "incoming patient_number_lv1 is null";
+        }
+
+        List<Integer> listPatientNumber = findAllPatientNumber();
+        if (!listPatientNumber.contains(patientLv1.getPatient_number_lv1())){
+            return "This patient doesn't exist";
         }
 
         patientDao.updatePatientLv1Info(patientLv1);
-        System.out.println("病人一級表更新記錄成功");
+        return "A recording in patient_lv1 is updated successfully";
     }
 
     @Override
-    public void insertPatientLv1Info(PatientLv1 patientLv1) {
+    public String insertPatientLv1Info(PatientLv1 patientLv1) {
         if (patientLv1.getPatient_number_lv1() == 0 || patientLv1.getName() == null){
-            System.out.println("病人一級表新增記錄時未傳入patient_number_lv1或name");
-            return;
+            return "incoming patient_number_lv1 or name is null";
         }
 
         patientDao.insertPatientLv1Info(patientLv1);
         System.out.println("病人一級表新增記錄成功");
+        return "A recording in patient_lv1 is inserted successfully";
     }
 
     @Override
-    public void updatePatientLv2Info(PatientLv2 patientLv2) {
+    public String updatePatientLv2Info(PatientLv2 patientLv2) {
         if (patientLv2.getPatient_number_lv2() == 0 || patientLv2.getRepresent() == null){
-            System.out.println("病人二級表更新記錄時未傳入patient_number_lv2或represent");
-            return;
+            return "incoming patient_number_lv2 or represent is null";
+        }
+
+        List<Integer> listPatientNumber = findAllPatientNumber();
+        if (!listPatientNumber.contains(patientLv2.getPatient_number_lv2())){
+            return "This patient doesn't exist";
         }
 
         patientDao.updatePatientLv2Info(patientLv2);
-        System.out.println("病人二級表更新記錄成功");
+        return "A recording in patient_lv2 is updated successfully";
     }
 
     @Override
-    public void insertPatientLv2Info(PatientLv2 patientLv2) {
+    public String insertPatientLv2Info(PatientLv2 patientLv2) {
         if (patientLv2.getPatient_number_lv2() == 0 || patientLv2.getRepresent() == null
             || patientLv2.getCheckbox_binary() == null){
-            System.out.println("病人二級表新增記錄時未傳入patient_number_lv2,represent或checkbox_binary");
-            return;
+            return "patient_number_lv2,represent or checkbox_binary is null";
         }
 
         patientDao.insertPatientLv2Info(patientLv2);
-        System.out.println("病人二級表新增記錄成功");
+        return "The record is inserted successfully in patient_lv2 ";
     }
 
     @Override
-    public void deletePatientLv2Info(PatientLv2 patientLv2) {
+    public String deletePatientLv2Info(PatientLv2 patientLv2) {
         if (patientLv2.getPatient_number_lv2() == 0 || patientLv2.getRepresent() == null){
-            System.out.println("病人二級表刪除記錄時未傳入patient_number_lv2或represent");
-            return;
+            return "incoming patient_number_lv2 or represent is null";
+        }
+
+        List<Integer> listPatientNumber = findAllPatientNumber();
+        if (!listPatientNumber.contains(patientLv2.getPatient_number_lv2())){
+            return "This patient doesn't exist";
         }
 
         patientDao.deletePatientLv2Info(patientLv2);
-        System.out.println("病人二級表刪除記錄成功");
+        return "The record is deleted successfully from patient_lv2";
     }
 
     @Override
-    public void deletePatientLv2InfoWithPatientNumber(PatientLv2 patientLv2) {
+    public String deletePatientLv2InfoWithPatientNumber(PatientLv2 patientLv2) {
         if (patientLv2.getPatient_number_lv2() == 0){
             System.out.println("病人二級表按住院號刪除時未傳入patient_number_lv2");
-            return;
+            return "incoming patient_number_lv2 is null";
+        }
+
+        List<Integer> listPatientNumber = findAllPatientNumber();
+        if (!listPatientNumber.contains(patientLv2.getPatient_number_lv2())){
+            return "This patient doesn't exist";
         }
 
         patientDao.deletePatientLv2InfoWithPatientNumber(patientLv2);
         System.out.println("病人二級表按住院號刪除成功");
+        return "The recordings are successfully deleted from patient_lv2 according to the patient number";
     }
 }
